requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25

# OCR and Image Processing
paddleocr==2.7.0.3
paddlepaddle==2.6.0
opencv-python==4.9.0.80
pdf2image==1.17.0
Pillow==10.2.0

# LLM Integration
requests==2.31.0
httpx==0.26.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
